requirements.txt
click
Flask
Flask-Cors
Flask-SQLAlchemy
greenlet
gunicorn
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
six
SQLAlchemy
Werkzeug
zipp
psycopg2-binary
flask_httpauth
pillow
werkzeug
glob2
flask-migrate
vonage

